Fix bugged dictionary clone/move when running in some docker configurations
When running with a mounted data directory, using shutil to move files would somehow cause the host OS  to see the move, but the process in the running container saw a stale/desynced filesystem. This caused a broken dictionary, as the database would point to the new location, while the file move wasn't visible. I tried many things to fix this, using named volumes, symlinks, running in privileged mode, manual sync() calls, but in the end the only "fix" that would actually work is using a copy+delete instead of a move.