Fix potential database locked error when quickly saving configs
After saving a config, the interface silently reloads the configs and entrylist, but since loading the entrylist can perform some internal checks when the dictionary needs to be reindexed, this can take a while. Attempting to (re-)save the config during the period would cause a database locked error, causing the backend to leak a connection somewhere, rendering the dictionary unusable until Lexonomy is restarted. Proper connection management should be implemented, but for now, this sidesteps the worst offender.